+++ test/aqua/test_initial_state_custom.py
@@ -12,92 +12,107 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
+""" Test Initial State Custom """
+
 import unittest
-
+from test.aqua.common import QiskitAquaTestCase
 import numpy as np
-
 from qiskit.aqua import AquaError
 from qiskit.aqua.components.initial_states import Custom
-from test.aqua.common import QiskitAquaTestCase
 
 
 class TestInitialStateCustom(QiskitAquaTestCase):
-
+    """ Test Initial State Custom """
     def test_qubits_2_zero_vector(self):
-        self.custom = Custom(2, state='zero')
-        cct = self.custom.construct_circuit('vector')
+        """ qubites 2 zero vector test """
+        custom = Custom(2, state='zero')
+        cct = custom.construct_circuit('vector')
         np.testing.assert_array_equal(cct, [1.0, 0.0, 0.0, 0.0])
 
     def test_qubits_5_zero_vector(self):
-        self.custom = Custom(5, state='zero')
-        cct = self.custom.construct_circuit('vector')
+        """ qubits 5 zero vector test """
+        custom = Custom(5, state='zero')
+        cct = custom.construct_circuit('vector')
         np.testing.assert_array_equal(cct, [1.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0,
                                             0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0,
                                             0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0,
                                             0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0])
 
     def test_qubits_2_zero_circuit(self):
-        self.custom = Custom(2, state='zero')
-        cct = self.custom.construct_circuit('circuit')
+        """ qubits 2 zero circuit test """
+        custom = Custom(2, state='zero')
+        cct = custom.construct_circuit('circuit')
         self.assertEqual(cct.qasm(), 'OPENQASM 2.0;\ninclude "qelib1.inc";\nqreg q[2];\n')
 
     def test_qubits_5_zero_circuit(self):
-        self.custom = Custom(5, state='zero')
-        cct = self.custom.construct_circuit('circuit')
+        """ qubits 5 zero circuit test """
+        custom = Custom(5, state='zero')
+        cct = custom.construct_circuit('circuit')
         self.assertEqual(cct.qasm(), 'OPENQASM 2.0;\ninclude "qelib1.inc";\nqreg q[5];\n')
 
     def test_qubits_2_uniform_vector(self):
-        self.custom = Custom(2, state='uniform')
-        cct = self.custom.construct_circuit('vector')
+        """ qubits 2 uniform vector test """
+        custom = Custom(2, state='uniform')
+        cct = custom.construct_circuit('vector')
         np.testing.assert_array_equal(cct, [0.5]*4)
 
     def test_qubits_5_uniform_vector(self):
-        self.custom = Custom(5, state='uniform')
-        cct = self.custom.construct_circuit('vector')
+        """ qubits 5 uniform vector test """
+        custom = Custom(5, state='uniform')
+        cct = custom.construct_circuit('vector')
         np.testing.assert_array_almost_equal(cct, [0.1767767]*32)
 
     def test_qubits_2_uniform_circuit(self):
-        self.custom = Custom(2, state='uniform')
-        cct = self.custom.construct_circuit('circuit')
-        self.assertEqual(cct.qasm(), 'OPENQASM 2.0;\ninclude "qelib1.inc";\nqreg q[2];\n'
-                                     'u2(0.0,3.14159265358979) q[0];\nu2(0.0,3.14159265358979) q[1];\n')
+        """ qubits 2 uniform circuit test """
+        custom = Custom(2, state='uniform')
+        cct = custom.construct_circuit('circuit')
+        self.assertEqual(cct.qasm(),
+                         'OPENQASM 2.0;\ninclude "qelib1.inc";\nqreg q[2];\n'
+                         'u2(0.0,3.14159265358979) q[0];\nu2(0.0,3.14159265358979) q[1];\n')
 
     def test_qubits_2_random_vector(self):
-        self.custom = Custom(2, state='random')
-        cct = self.custom.construct_circuit('vector')
+        """ qubits 2 random vector test """
+        custom = Custom(2, state='random')
+        cct = custom.construct_circuit('vector')
         prob = np.sqrt(np.sum([x**2 for x in cct]))
         self.assertAlmostEqual(prob, 1.0)
 
     def test_qubits_5_random_vector(self):
-        self.custom = Custom(5, state='random')
-        cct = self.custom.construct_circuit('vector')
+        """ qubits 5 random vector test """
+        custom = Custom(5, state='random')
+        cct = custom.construct_circuit('vector')
         prob = np.sqrt(np.sum([x**2 for x in cct]))
         self.assertAlmostEqual(prob, 1.0)
 
     def test_qubits_2_given_vector(self):
-        self.custom = Custom(2, state_vector=[0.5]*4)
-        cct = self.custom.construct_circuit('vector')
+        """ qubits 2 given vector test """
+        custom = Custom(2, state_vector=[0.5]*4)
+        cct = custom.construct_circuit('vector')
         np.testing.assert_array_equal(cct, [0.5]*4)
 
     def test_qubits_5_given_vector(self):
-        self.custom = Custom(5, state_vector=[1.0]*32)
-        cct = self.custom.construct_circuit('vector')
+        """ qubits 5 given vector test """
+        custom = Custom(5, state_vector=[1.0]*32)
+        cct = custom.construct_circuit('vector')
         np.testing.assert_array_almost_equal(cct, [0.1767767]*32)
 
     def test_qubits_5_randgiven_vector(self):
-        self.custom = Custom(5, state_vector=np.random.rand(32))
-        cct = self.custom.construct_circuit('vector')
+        """ qubits 5 randgiven vector test """
+        custom = Custom(5, state_vector=np.random.rand(32))
+        cct = custom.construct_circuit('vector')
         prob = np.sqrt(np.sum([x**2 for x in cct]))
         self.assertAlmostEqual(prob, 1.0)
 
-    def test_qubits_qubits_given_mistmatch(self):
+    def test_qubits_qubits_given_mismatch(self):
+        """ qubits 5 given mismatch test """
         with self.assertRaises(AquaError):
-            self.custom = Custom(5, state_vector=[1.0]*23)
+            _ = Custom(5, state_vector=[1.0]*23)
 
     def test_qubits_2_zero_vector_wrong_cct_mode(self):
-        self.custom = Custom(5, state='zero')
+        """ qubits 2 zero vector wrong cct mode test """
+        custom = Custom(5, state='zero')
         with self.assertRaises(AquaError):
-            cct = self.custom.construct_circuit('matrix')
+            _ = custom.construct_circuit('matrix')
 
 
 if __name__ == '__main__':